Move Entity to Framework
