+++ src/Core/Connector/ConnectionReader.php
@@ -2,7 +2,7 @@
 
 namespace Smolblog\Core\Connector;
 
-use Smolblog\Core\Entity\Reader;
+use Smolblog\Framework\Reader;
 
 /**
  * An object that can retrieve an Connection from a repository.
