+++ tests/Core/Command/CommandBusTest.php
