+++ tests/Core/Connector/AuthRequestFinalizerTest.php
