+++ tests/Core/Connector/AuthRequestInitializerTest.php
