+++ tests/Core/Connector/ChannelRefresherTest.php
