+++ tests/Core/Connector/ConnectCallbackTest.php
