+++ tests/Core/Connector/ConnectInitTest.php
