+++ tests/Core/Connector/ConnectionTest.php
