+++ tests/Core/Connector/ConnectorRegistrarTest.php
