+++ tests/Core/Connector/RefreshConnectionTokenTest.php
