+++ tests/Core/Connector/UserConnectionsTest.php
