+++ tests/Core/EnvironmentTest.php
