+++ tests/Core/Events/EventDispatcherTest.php
