+++ tests/Core/Importer/ImporterRegistrarTest.php
