+++ tests/Core/Plugin/InstalledPluginsTest.php
