+++ tests/Core/Registrar/RegistrarTest.php
