--- tests/_pluginStub/src/Plugin.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Smoltest\PluginStub;
-
-use Smolblog\Core\App;
-use Smolblog\Core\Plugin\Plugin as SmolblogPlugin;
-
-class Plugin implements SmolblogPlugin {
-	public function __construct(App $smolblog) {}
-}
